product attribute fixed

+++ app/Http/Controllers/ProductAttributeController.php
@@ -28,7 +28,7 @@ class ProductAttributeController extends Controller
         return Inertia::render('Admin/ProductAttribute/Index', [
             'attributes' => $product_attributes
         ])->table(function (InertiaTable $table) {
-            $table->withGlobalSearch()
+            $table
                 ->column('id', sortable: true)
                 ->column('name', sortable: true, searchable: true)
                 ->column('type')
@@ -55,6 +55,7 @@ class ProductAttributeController extends Controller
         $product_attribute = new Attribute();
         $product_attribute->fill($request->validated());
         $product_attribute->save();
+        $product_attribute->add_options($request->validated()['options']);
 
         return to_route('admin.attributes.index')->with('success', 'Product Attribute created successfully.');
     }
@@ -82,7 +83,15 @@ class ProductAttributeController extends Controller
      */
     public function update(UpdateProductAttributeRequest $request, Attribute $attribute): RedirectResponse
     {
+
         $attribute->fill($request->validated());
+
+        if ($request->has('options')){
+            $attribute->options()->delete();
+            $attribute->add_options($request->validated()['options']);
+        }
+
+//        dd($request->validated());
         $attribute->save();
 
         return back()->with('success', 'Product Attribute updated successfully.');
@@ -93,6 +102,9 @@ class ProductAttributeController extends Controller
      */
     public function destroy(Attribute $attribute): RedirectResponse
     {
+        if ($attribute->products()->count()){
+            return to_route('admin.attributes.index')->with('error', 'Product Attribute can not be deleted!.');
+        }
         $attribute->delete();
         return to_route('admin.attributes.index')->with('success', 'Product Attribute deleted successfully.');
     }
